Extract `admin_move_market_to_closed` tests
