+++ zrml/prediction-markets/src/old_tests.rs
@@ -141,150 +141,6 @@ fn simple_create_scalar_market(
     ));
 }
 
-#[test]
-fn admin_move_market_to_closed_successfully_closes_market_and_sets_end_blocknumber() {
-    ExtBuilder::default().build().execute_with(|| {
-        run_blocks(7);
-        let now = frame_system::Pallet::<Runtime>::block_number();
-        let end = 42;
-        simple_create_categorical_market(
-            Asset::Ztg,
-            MarketCreation::Permissionless,
-            now..end,
-            ScoringRule::Lmsr,
-        );
-        run_blocks(3);
-        let market_id = 0;
-        assert_ok!(PredictionMarkets::admin_move_market_to_closed(
-            RuntimeOrigin::signed(SUDO),
-            market_id
-        ));
-        let market = MarketCommons::market(&market_id).unwrap();
-        assert_eq!(market.status, MarketStatus::Closed);
-        let new_end = now + 3;
-        assert_eq!(market.period, MarketPeriod::Block(now..new_end));
-        assert_ne!(new_end, end);
-        System::assert_last_event(Event::MarketClosed(market_id).into());
-    });
-}
-
-#[test]
-fn admin_move_market_to_closed_successfully_closes_market_and_sets_end_timestamp() {
-    ExtBuilder::default().build().execute_with(|| {
-        let start_block = 7;
-        set_timestamp_for_on_initialize(start_block * MILLISECS_PER_BLOCK as u64);
-        run_blocks(start_block);
-        let start = <zrml_market_commons::Pallet<Runtime>>::now();
-
-        let end = start + 42.saturated_into::<TimeFrame>() * MILLISECS_PER_BLOCK as u64;
-        assert_ok!(PredictionMarkets::create_market(
-            RuntimeOrigin::signed(ALICE),
-            Asset::Ztg,
-            Perbill::zero(),
-            BOB,
-            MarketPeriod::Timestamp(start..end),
-            get_deadlines(),
-            gen_metadata(2),
-            MarketCreation::Permissionless,
-            MarketType::Categorical(<Runtime as crate::Config>::MinCategories::get()),
-            Some(MarketDisputeMechanism::SimpleDisputes),
-            ScoringRule::Lmsr
-        ));
-        let market_id = 0;
-        let market = MarketCommons::market(&market_id).unwrap();
-        assert_eq!(market.period, MarketPeriod::Timestamp(start..end));
-
-        let shift_blocks = 3;
-        let shift = shift_blocks * MILLISECS_PER_BLOCK as u64;
-        // millisecs per block is substracted inside the function
-        set_timestamp_for_on_initialize(start + shift + MILLISECS_PER_BLOCK as u64);
-        run_blocks(shift_blocks);
-
-        assert_ok!(PredictionMarkets::admin_move_market_to_closed(
-            RuntimeOrigin::signed(SUDO),
-            market_id
-        ));
-        let market = MarketCommons::market(&market_id).unwrap();
-        assert_eq!(market.status, MarketStatus::Closed);
-        let new_end = start + shift;
-        assert_eq!(market.period, MarketPeriod::Timestamp(start..new_end));
-        assert_ne!(new_end, end);
-        System::assert_last_event(Event::MarketClosed(market_id).into());
-    });
-}
-
-#[test]
-fn admin_move_market_to_closed_fails_if_market_does_not_exist() {
-    ExtBuilder::default().build().execute_with(|| {
-        assert_noop!(
-            PredictionMarkets::admin_move_market_to_closed(RuntimeOrigin::signed(SUDO), 0),
-            zrml_market_commons::Error::<Runtime>::MarketDoesNotExist
-        );
-    });
-}
-
-#[test_case(MarketStatus::Closed; "closed")]
-#[test_case(MarketStatus::Reported; "reported")]
-#[test_case(MarketStatus::Disputed; "disputed")]
-#[test_case(MarketStatus::Resolved; "resolved")]
-#[test_case(MarketStatus::Proposed; "proposed")]
-fn admin_move_market_to_closed_fails_if_market_is_not_active(market_status: MarketStatus) {
-    ExtBuilder::default().build().execute_with(|| {
-        simple_create_categorical_market(
-            Asset::Ztg,
-            MarketCreation::Permissionless,
-            0..1,
-            ScoringRule::Lmsr,
-        );
-        let market_id = 0;
-        assert_ok!(MarketCommons::mutate_market(&market_id, |market| {
-            market.status = market_status;
-            Ok(())
-        }));
-        assert_noop!(
-            PredictionMarkets::admin_move_market_to_closed(RuntimeOrigin::signed(SUDO), market_id),
-            Error::<Runtime>::MarketIsNotActive,
-        );
-    });
-}
-
-#[test]
-fn admin_move_market_to_closed_correctly_clears_auto_close_blocks() {
-    ExtBuilder::default().build().execute_with(|| {
-        let category_count = 3;
-        assert_ok!(PredictionMarkets::create_market(
-            RuntimeOrigin::signed(ALICE),
-            Asset::Ztg,
-            Perbill::zero(),
-            ALICE,
-            MarketPeriod::Block(22..66),
-            get_deadlines(),
-            gen_metadata(50),
-            MarketCreation::Permissionless,
-            MarketType::Categorical(category_count),
-            Some(MarketDisputeMechanism::SimpleDisputes),
-            ScoringRule::Lmsr,
-        ));
-        assert_ok!(PredictionMarkets::create_market(
-            RuntimeOrigin::signed(ALICE),
-            Asset::Ztg,
-            Perbill::zero(),
-            ALICE,
-            MarketPeriod::Block(33..66),
-            get_deadlines(),
-            gen_metadata(50),
-            MarketCreation::Permissionless,
-            MarketType::Categorical(category_count),
-            Some(MarketDisputeMechanism::SimpleDisputes),
-            ScoringRule::Lmsr,
-        ));
-        assert_ok!(PredictionMarkets::admin_move_market_to_closed(RuntimeOrigin::signed(SUDO), 0));
-
-        let auto_close = MarketIdsPerCloseBlock::<Runtime>::get(66).into_inner();
-        assert_eq!(auto_close, vec![1]);
-    });
-}
-
 #[test_case(654..=321; "empty range")]
 #[test_case(555..=555; "one element as range")]
 fn create_scalar_market_fails_on_invalid_range(range: RangeInclusive<u128>) {
